Updated labelling code and labels (see note)
The version of 05-labelling/02-labelling.py used for labelling included a condition for (meanwj['Name'].isna()) in line 38. As such, for each feature, the code would return the top 20 unlabelled topics. This means that extra topics were labelled beyond the top 20 in each feature (since some topics appear in the true top 20 across several lists). The extra topic labels are preserved for transparency, but are not included in further analysis since they were dependent on the ordering of the features. The extra labels were identified by re-sorting the dataframe on each feature and collecting the topic IDs that did not appear in any of the top 20s.

All labels recorded in file: support_data/label agreement (extra labels).xlsx
Labels used for presentation and accuracy calculations going forward are in
file: support_data/label agreement.xlsx